translation-service/requirements.txt
# OPUS-MT Translation Model
transformers>=4.36.0
sentencepiece>=0.1.99
sacremoses>=0.0.53

# For CPU-only (no CUDA)
torch>=2.0.0

# PDF/DOC Processing
PyMuPDF>=1.23.0
python-docx>=1.1.0

# HTTP & JSON
requests>=2.31.0
aiohttp>=3.9.0

# FastAPI (for API mode)
fastapi>=0.109.0
uvicorn>=0.27.0

# RabbitMQ
pika>=1.3.2

# PostgreSQL
psycopg2-binary>=2.9.9

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

